Replace int value of Feature by double.

+++ features-extraction/src/main/java/org/jetbrains/research/groups/ml_methods/extraction/features/AnotherInstanceCallers.java
@@ -8,7 +8,7 @@ package org.jetbrains.research.groups.ml_methods.extraction.features;
  * somewhere else.
  */
 public class AnotherInstanceCallers extends Feature {
-    public AnotherInstanceCallers(int value) {
+    public AnotherInstanceCallers(double value) {
         super(value);
     }
 }
