+++ features-extraction/src/main/java/org/jetbrains/research/groups/ml_methods/extraction/features/TargetClassStaticCallTargets.java
@@ -5,7 +5,7 @@ package org.jetbrains.research.groups.ml_methods.extraction.features;
  * {@link TargetClassInstanceCallTargets} but it counts static methods.
  */
 public class TargetClassStaticCallTargets extends Feature {
-    public TargetClassStaticCallTargets(int value) {
+    public TargetClassStaticCallTargets(double value) {
         super(value);
     }
 }
